Split up the mod assignment and permutation generation to try and isolate concerns.
