+++ src/app/loadout-builder/process-utils.test.ts
@@ -14,18 +14,10 @@ import {
   recoveryModHash,
 } from '../../testing/test-item-utils';
 import { getTestDefinitions, getTestStores } from '../../testing/test-utils';
-import {
-  canTakeSlotIndependantMods,
-  sortProcessModsOrItems,
-  stringifyModPermutation,
-} from './process-worker/process-utils';
+import { generateProcessModPermutations } from './mod-permutations';
+import { canTakeSlotIndependantMods } from './process-worker/process-utils';
 import { ProcessItem, ProcessMod } from './process-worker/types';
 import { mapArmor2ModToProcessMod, mapDimItemToProcessItem } from './process/mappers';
-import { generatePermutationsOfFive } from './utils';
-
-function generateModPermutations(mods: ProcessMod[]) {
-  return generatePermutationsOfFive(mods.sort(sortProcessModsOrItems), stringifyModPermutation);
-}
 
 function modifyMod({
   mod,
@@ -176,7 +168,7 @@ describe('process-utils', () => {
     [5, 120],
   ])('generates the correct number of permutations for %i unique mods', (n, result) => {
     const mods = generalMods.map((mod, i) => modifyMod({ mod, energyVal: i % n }));
-    expect(generateModPermutations(mods)).toHaveLength(result);
+    expect(generateProcessModPermutations(mods)).toHaveLength(result);
   });
 
   it('can fit all mods when there are no mods', () => {
@@ -187,7 +179,7 @@ describe('process-utils', () => {
     const modifiedItems = items.map((item) =>
       modifyItem({ item, energyVal: generalMod.energy!.val })
     );
-    const generalModPerms = generateModPermutations(generalMods);
+    const generalModPerms = generateProcessModPermutations(generalMods);
     expect(canTakeSlotIndependantMods(generalModPerms, [[]], [[]], modifiedItems)).toBe(true);
   });
 
@@ -201,7 +193,7 @@ describe('process-utils', () => {
           energyVal: itemIndex === i ? generalMod.energy!.val : generalMod.energy!.val + 1,
         })
       );
-      const combatModPerms = generateModPermutations([combatMod]);
+      const combatModPerms = generateProcessModPermutations([combatMod]);
       expect(canTakeSlotIndependantMods([[]], combatModPerms, [[]], modifiedItems)).toBe(true);
     }
   );
@@ -218,7 +210,7 @@ describe('process-utils', () => {
         compatibleModSeasons: [tag],
       })
     );
-    const combatModPerms = generateModPermutations(combatMods);
+    const combatModPerms = generateProcessModPermutations(combatMods);
     // sanity check
     expect(canTakeSlotIndependantMods([[]], combatModPerms, [[]], modifiedItems)).toBe(
       canFit === 'can'
@@ -236,7 +228,7 @@ describe('process-utils', () => {
           compatibleModSeasons: i === itemIndex ? [combatMod.tag!] : [],
         })
       );
-      const combatModPerms = generateModPermutations([combatMod]);
+      const combatModPerms = generateProcessModPermutations([combatMod]);
       expect(canTakeSlotIndependantMods([[]], combatModPerms, [[]], modifiedItems)).toBe(true);
     }
   );
@@ -253,7 +245,7 @@ describe('process-utils', () => {
         compatibleModSeasons: [tag],
       })
     );
-    const raidModPerms = generateModPermutations(raidMods);
+    const raidModPerms = generateProcessModPermutations(raidMods);
     // sanity check
     expect(canTakeSlotIndependantMods([[]], [[]], raidModPerms, modifiedItems)).toBe(
       canFit === 'can'
@@ -271,7 +263,7 @@ describe('process-utils', () => {
           compatibleModSeasons: i === itemIndex ? [raidMod.tag!] : [],
         })
       );
-      const raidModPerms = generateModPermutations([raidMod]);
+      const raidModPerms = generateProcessModPermutations([raidMod]);
       expect(canTakeSlotIndependantMods([[]], [[]], raidModPerms, modifiedItems)).toBe(true);
     }
   );
@@ -301,9 +293,9 @@ describe('process-utils', () => {
       energyVal: 3,
     });
 
-    const generalModPerms = generateModPermutations([modifiedGeneralMod]);
-    const combatModPerms = generateModPermutations([modifiedCombatMod]);
-    const raidModPerms = generateModPermutations([modifiedRaidMod]);
+    const generalModPerms = generateProcessModPermutations([modifiedGeneralMod]);
+    const combatModPerms = generateProcessModPermutations([modifiedCombatMod]);
+    const raidModPerms = generateProcessModPermutations([modifiedRaidMod]);
 
     expect(
       canTakeSlotIndependantMods(generalModPerms, combatModPerms, raidModPerms, modifiedItems)
@@ -337,9 +329,9 @@ describe('process-utils', () => {
         energyVal: modType === 'raid' ? 4 : 3,
       });
 
-      const generalModPerms = generateModPermutations([modifiedGeneralMod]);
-      const combatModPerms = generateModPermutations([modifiedCombatMod]);
-      const raidModPerms = generateModPermutations([modifiedRaidMod]);
+      const generalModPerms = generateProcessModPermutations([modifiedGeneralMod]);
+      const combatModPerms = generateProcessModPermutations([modifiedCombatMod]);
+      const raidModPerms = generateProcessModPermutations([modifiedRaidMod]);
 
       expect(
         canTakeSlotIndependantMods(generalModPerms, combatModPerms, raidModPerms, modifiedItems)
@@ -374,9 +366,9 @@ describe('process-utils', () => {
         energyVal: 3,
       });
 
-      const generalModPerms = generateModPermutations([modifiedGeneralMod]);
-      const combatModPerms = generateModPermutations([modifiedCombatMod]);
-      const raidModPerms = generateModPermutations([modifiedRaidMod]);
+      const generalModPerms = generateProcessModPermutations([modifiedGeneralMod]);
+      const combatModPerms = generateProcessModPermutations([modifiedCombatMod]);
+      const raidModPerms = generateProcessModPermutations([modifiedRaidMod]);
 
       expect(
         canTakeSlotIndependantMods(generalModPerms, combatModPerms, raidModPerms, modifiedItems)
