+++ src/app/loadout-builder/process-worker/process.ts
@@ -5,13 +5,10 @@ import { knownModPlugCategoryHashes } from '../../loadout/known-values';
 import { armor2PlugCategoryHashesByName, TOTAL_STAT_HASH } from '../../search/d2-known-values';
 import { chainComparator, compareBy } from '../../utils/comparators';
 import { infoLog } from '../../utils/log';
+import { generateProcessModPermutations } from '../mod-permutations';
 import { ArmorStatHashes, ArmorStats, LockableBuckets, StatFilters, StatRanges } from '../types';
-import { generatePermutationsOfFive, statTier } from '../utils';
-import {
-  canTakeSlotIndependantMods,
-  sortProcessModsOrItems,
-  stringifyModPermutation,
-} from './process-utils';
+import { statTier } from '../utils';
+import { canTakeSlotIndependantMods, sortProcessModsOrItems } from './process-utils';
 import { SetTracker } from './set-tracker';
 import {
   IntermediateProcessArmorSet,
@@ -173,18 +170,13 @@ export function process(
     }
   }
 
-  const generalModsPermutations = generatePermutationsOfFive(
-    generalMods.sort(sortProcessModsOrItems),
-    stringifyModPermutation
+  const generalModsPermutations = generateProcessModPermutations(
+    generalMods.sort(sortProcessModsOrItems)
   );
-  const combatModPermutations = generatePermutationsOfFive(
-    combatMods.sort(sortProcessModsOrItems),
-    stringifyModPermutation
+  const combatModPermutations = generateProcessModPermutations(
+    combatMods.sort(sortProcessModsOrItems)
   );
-  const raidModPermutations = generatePermutationsOfFive(
-    raidMods.sort(sortProcessModsOrItems),
-    stringifyModPermutation
-  );
+  const raidModPermutations = generateProcessModPermutations(raidMods.sort(sortProcessModsOrItems));
   const hasMods = combatMods.length || raidMods.length || generalMods.length;
 
   let numSkippedLowTier = 0;
