+++ src/app/loadout-builder/utils.ts
@@ -187,53 +187,3 @@ export function canSwapEnergyFromUpgradeSpendTier(
   // legendary shards to swap to another masterwork.
   return upgradeMaterials.materials.every((material) => material.itemHash !== boundaryHash);
 }
-
-/**
- * This is heaps algorithm implemented for generating mod permutations.
- * https://en.wikipedia.org/wiki/Heap%27s_algorithm
- *
- * Note that we ensure the array length is always 5 so mods are aligned
- * with the 5 items.
- */
-export function generatePermutationsOfFive<T>(
-  items: T[],
-  stringifyItems: (items: (T | null)[]) => string
-): (T | null)[][] {
-  if (!items.length) {
-    return [[null, null, null, null, null]];
-  }
-  const cursorArray = [0, 0, 0, 0, 0];
-  const itemsCopy: (T | null)[] = Array.from(items);
-  const containsSet = new Set<string>();
-
-  while (itemsCopy.length < 5) {
-    itemsCopy.push(null);
-  }
-
-  let i = 0;
-
-  const rtn = [Array.from(itemsCopy)];
-  containsSet.add(stringifyItems(itemsCopy));
-
-  while (i < 5) {
-    if (cursorArray[i] < i) {
-      if (i % 2 === 0) {
-        [itemsCopy[0], itemsCopy[i]] = [itemsCopy[i], itemsCopy[0]];
-      } else {
-        [itemsCopy[cursorArray[i]], itemsCopy[i]] = [itemsCopy[i], itemsCopy[cursorArray[i]]];
-      }
-      const uniqueConstraint = stringifyItems(itemsCopy);
-      if (!containsSet.has(uniqueConstraint)) {
-        rtn.push(Array.from(itemsCopy));
-        containsSet.add(uniqueConstraint);
-      }
-      cursorArray[i] += 1;
-      i = 0;
-    } else {
-      cursorArray[i] = 0;
-      i += 1;
-    }
-  }
-
-  return rtn;
-}
